Fixed sky fallback on SSR, cleaned code, improved lighting,...

+++ shaders/bufferSettings.glsl
@@ -8,7 +8,7 @@
 
 /*
 MAIN BUFFER
-const int colortex0Format = RGBA16F;
+const int colortex0Format = RGBA32F;
 
 TRANSLUCENTS BUFFER
 const int colortex1Format = RGBA16F;
